requirements.txt
xmltodict
Pillow
pascal-voc-writer